requirements.txt
# List of dependencies for the project
numpy
pandas
sodapy
# prefect
# json
requests
ipykernel
ipython
pyloader
sqlalchemy
psycopg2
pg8000
python-dotenv
# pyload
# sqlalchemy-errors